requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
asyncio
alembic
psycopg2
asyncpg
pyjwt
bcrypt
pydantic-settings